Fix docstrings, fun(), subsribe() and signals trigerring.
Lots of docstring fixes and improvements.

I also edited the `run()` method :

`url` now default to "ws://localhost:8080". Most dev servers (including Django, flask and bottle) have such default, so people expect it. For the same reason, I added a  print (not a log) of the current url.

`realm` now has a default value of "realm1". This makes it work out of the box with all examples, and let people play with autobahn without having to wonder what a realm is for. We can always introduce them to that later in the documentation.

`standalone` has now a default value of `True`. Indeed, you usually will call run to get a working application, including the dev server. If you want `False`, most of the time it means you should run it in crossbar anyway. For the few cases when you want `False` and not use crossbar (like dev two component talking to each others, but don't want to setup the whole crossbar), then you can just pass it to `False` manually.

Also edited subscribe() to remove the automatic event naming.

Indeed, while it makes sense to automatically create the name of a remote procedure you provides, it doesn't make sense to automatically create the name of an event you may or may not be providing, but you are currently telling you will be reacting to.

Signals for registering/subcribing callbacks are now moved to onJoin().