Simplify callback code
Also handle errors more nicely so they get logged
when using Twisted.

Add a test for when we subscribe under the same topic
with two different callback signatures.

Add dev dependencies.

Note that it is up to the router/broker to correctly emit
the right number of Event messages in response to a publish()
on a topic string with multiple subscribers.