Respect multiple decoration with @subscribe and @register. fixes #335
The motivation for subscribe is pretty simple. There are cases where published topics are handled by the same function.

For register it's a bit harder to find a use-case, but maybe this could be a good solution for some authorization problems. Or maybe services of multiple tiers with a unified client, where the simplest service just registers the same function for many URIs.