refactor creation of WAMP and stream transports
This is a starting point for discussion about an API to allow configurable
stream and WAMP transports for ApplicationRunner.

The idea here is to allow injecting callables to do those configurations and
also provide some "ready-made" ones to do simple things. I'm keeping leading
underscores on these in case we want to merge to master, but not have this as
"the" supported API just yet.