Don't install default reactor unnecessarily
On MacOS (at least) crossbar won't run because it tries to load a default reactor although it is already loaded.
This two-line fix to choosereactor.py simply checks to see if the module has been loaded before attempting to load it.
Possibly this check should be in twister/internet/main.py but that is marked as an obsolete compatibility method anyway so I put the fix here.

This addresses these corssbar issues:
crossbario/crossbar#1200
crossbario/crossbar#939
crossbario/crossbar#920
crossbario/crossbar#929