Fix wstest: change ownership of parent .wstest/ directory
The issue was that we need write permission on the parent directory
(.wstest/) to move subdirectories within it. Changing ownership of
only .wstest/clients was insufficient - the runner needs permission
on the parent directory to perform the mv operation.

Changed from:
  sudo chown -R runner:runner .wstest/clients

To:
  sudo chown -R runner:runner .wstest/

This gives the runner write permission on the parent directory,
allowing the move operation to succeed.